Create MenuAssignmentBash
The while loop will continuously display the menu until the user chooses to exit the script by selecting option 6.

Option 1 uses the ls command to list all files in the current directory.

Option 2 uses the df -h command to show the free disk space available.

Option 3 uses the pwd command to show the current path.

Option 4 uses the history command to show the command history.

Option 5 prompts the user to enter the directory they want to backup and then creates a new directory called BackupFolder in the current directory to store the backup. The cp -R command is used to recursively copy the specified directory and its sub-directories to the BackupFolder directory. Finally, the contents of the BackupFolder directory are listed using the ls command.

Option 6 exits the script with an exit code of 0.

For each option, the user is prompted to press enter to continue before returning to the main menu.